pdb.h: Fix integer overflow in pdb_entry_pointer()
This is an accidental regression introduced in revision
802ef0eeff42c30159b04045f61b6bccbfd1b0a4 "pdb.c: Simplify PDB structure"
This regression wasn't spotted before as I rarely had a need to use
large pattern databases.

+++ pdb.h
@@ -108,9 +108,11 @@ static inline atomic_uchar *
 pdb_entry_pointer(struct patterndb *pdb, const struct index *idx)
 {
 	if (tileset_has(pdb->aux.ts, ZERO_TILE))
-		return (pdb->data + (pdb->aux.idxt[idx->maprank].offset + idx->eqidx) * pdb->aux.n_perm + idx->pidx);
+		return (pdb->data + (pdb->aux.idxt[idx->maprank].offset +
+		    idx->eqidx) * (size_t)pdb->aux.n_perm + idx->pidx);
 	else
-		return (pdb->data + idx->maprank * pdb->aux.n_perm + idx->pidx);
+		return (pdb->data + idx->maprank * (size_t)pdb->aux.n_perm +
+		    idx->pidx);
 }
 
 /*
